% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConfidenceIntervals.R
\name{estimateConfidenceIntervals}
\alias{estimateConfidenceIntervals}
\title{Estimate confidence intervals for Bland-Altman limits of agreement}
\usage{
estimateConfidenceIntervals(LOA, n, alpha = 0.05)
}
\arguments{
\item{LOA}{an object created by estimateLimitsOfAgreement}

\item{n}{sample size}

\item{alpha}{alpha level for calculating limits of agreement}
}
\value{
a blandaltman object
}
\description{
Estimate confidence intervals for Bland-Altman limits of agreement
}
