% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLimitsOfAgreement.R
\name{estimateLimitsOfAgreement}
\alias{estimateLimitsOfAgreement}
\title{Estimate Bland-Altman limits of agreement}
\usage{
estimateLimitsOfAgreement(mu, SD, gamma = 0.05)
}
\arguments{
\item{mu}{mean of the pairwise differences}

\item{SD}{standard deviation of the pairwise differences}

\item{gamma}{alpha level for calculating limits of agreement}
}
\value{
a list
}
\description{
Estimate Bland-Altman limits of agreement
}
