% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePower.R
\name{estimatePower}
\alias{estimatePower}
\title{Estimate power for Bland-Altman limits of agreement test}
\usage{
estimatePower(beta, method = "Lu")
}
\arguments{
\item{beta}{an object produced by estimateTypeIIerror}

\item{method}{whether to use the Lu or Wisniewski method for power}
}
\value{
a list
}
\description{
Estimate power for Bland-Altman limits of agreement test
}
