% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePowerCurve.R
\name{estimatePowerCurve}
\alias{estimatePowerCurve}
\title{Estimate power curve for Bland-Altman limits of agreement test}
\usage{
estimatePowerCurve(nMin = 10, nMax = 100, stepsize = 1, mu, SD,
  delta, gamma = 0.05, alpha = 0.05, approx = "t", method = "Lu")
}
\arguments{
\item{nMin}{minimum sample size}

\item{nMax}{maximum sample size}

\item{stepsize}{number of samples to step between nMin and nMax}

\item{mu}{mean of differences}

\item{SD}{standard deviation of differences}

\item{delta}{pre-determined clinical agreement interval}

\item{gamma}{alpha level of Bland-Altman LOA}

\item{alpha}{alpha level of LOA confidence intervals}

\item{approx}{normal or t distribution}

\item{method}{whether to use the Lu or Wisniewski method for power}
}
\value{
a dataframe
}
\description{
Estimate power curve for Bland-Altman limits of agreement test
}
