% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePowerFromPowerCurve.R
\name{estimatePowerFromPowerCurve}
\alias{estimatePowerFromPowerCurve}
\title{Estimate power for Bland-Altman limits of agreement test}
\usage{
estimatePowerFromPowerCurve(powerCurve, n)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}

\item{n}{the pre-determined sample size}
}
\value{
a list
}
\description{
Estimate power for Bland-Altman limits of agreement test
}
