% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSampleSize.R
\name{estimateSampleSize}
\alias{estimateSampleSize}
\title{Estimate sample size for Bland-Altman limits of agreement test}
\usage{
estimateSampleSize(powerCurve, power = 0.8)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}

\item{power}{the pre-determined power}
}
\value{
a list
}
\description{
Estimate sample size for Bland-Altman limits of agreement test
}
