% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSampleSize.R
\name{estimateSampleSize_mu0}
\alias{estimateSampleSize_mu0}
\title{Estimate sample size for Bland-Altman limits of agreement test assuming zero mean}
\usage{
estimateSampleSize_mu0(SD, delta, power = 0.8, gamma = 0.05,
  alpha = 0.05, iterMax = 100, tolerance = 0.001, method = "Lu",
  debug = FALSE)
}
\arguments{
\item{SD}{standard deviation of differences}

\item{delta}{pre-determined clinical agreement interval}

\item{power}{the pre-determined power}

\item{gamma}{alpha level of Bland-Altman LOA}

\item{alpha}{alpha level of LOA confidence intervals}

\item{iterMax}{maximum number of iterations}

\item{tolerance}{tolerance for how close the power estimate should be to the input}

\item{method}{whether to use the Lu or Wisniewski method for power}

\item{debug}{logical}
}
\value{
a sample size
}
\description{
Estimate sample size for Bland-Altman limits of agreement test assuming zero mean
}
