% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSampleSizefromPowerCurve.R
\name{estimateSampleSizeFromPowerCurve}
\alias{estimateSampleSizeFromPowerCurve}
\title{Estimate sample size for Bland-Altman limits of agreement test}
\usage{
estimateSampleSizeFromPowerCurve(powerCurve, power = 0.8,
  tolerance = 0.01)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}

\item{power}{the pre-determined power}

\item{tolerance}{how close the nearest estimated power must be to the pre-determined power}
}
\value{
a list
}
\description{
Estimate sample size for Bland-Altman limits of agreement test
}
