% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTypeIIerror.R
\name{estimateTypeIIerror}
\alias{estimateTypeIIerror}
\title{Estimate Type II error for Bland-Altman limits of agreement test}
\usage{
estimateTypeIIerror(CI, delta, approx = "t")
}
\arguments{
\item{CI}{an object created by estimateConfidenceIntervals}

\item{delta}{pre-determined clinical agreement level goal}

\item{approx}{whether to use the normal or student's t approximations}
}
\value{
a blandaltman object
}
\description{
Estimate Type II error for Bland-Altman limits of agreement test
}
