% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfidenceIntervalCurve.R
\name{plotConfidenceIntervalCurve}
\alias{plotConfidenceIntervalCurve}
\title{Plot confidence interval curves for Bland-Altman limits of agreement test}
\usage{
plotConfidenceIntervalCurve(powerCurve)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}
}
\value{
a ggplot2 object
}
\description{
Plot confidence interval curves for Bland-Altman limits of agreement test
}
