% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerCurve.R
\name{plotPowerCurve}
\alias{plotPowerCurve}
\title{Plot sample size vs power for Bland-Altman limits of agreement test}
\usage{
plotPowerCurve(powerCurve)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}
}
\value{
a ggplot2 object
}
\description{
Plot sample size vs power for Bland-Altman limits of agreement test
}
