% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePower.R
\name{simulatePower}
\alias{simulatePower}
\title{Simulate power for Bland-Altman limits of agreement test}
\usage{
simulatePower(mu, SD, delta, n, nsims = 1000, gamma = 0.05,
  alpha = 0.05)
}
\arguments{
\item{mu}{mean of differences}

\item{SD}{standard deviation of differences}

\item{delta}{pre-determined clinical agreement interval}

\item{n}{sample size}

\item{nsims}{number of simulations}

\item{gamma}{alpha level of Bland-Altman LOA}

\item{alpha}{alpha level of LOA confidence intervals}
}
\value{
a number expression the power
}
\description{
Simulate power for Bland-Altman limits of agreement test
}
