% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePowerCurve.R
\name{simulatePowerCurve}
\alias{simulatePowerCurve}
\title{Estimate power curve for Bland-Altman limits of agreement test}
\usage{
simulatePowerCurve(nMin, nMax, stepsize, mu, SD, delta, gamma = 0.05,
  alpha = 0.05, nsims = 1000, parallel = TRUE, ncores = NULL)
}
\arguments{
\item{nMin}{minimum sample size}

\item{nMax}{maximum sample size}

\item{stepsize}{number of samples to step between nMin and nMax}

\item{mu}{mean of differences}

\item{SD}{standard deviation of differences}

\item{delta}{pre-determined clinical agreement interval}

\item{gamma}{alpha level of Bland-Altman LOA}

\item{alpha}{alpha level of LOA confidence intervals}

\item{nsims}{number of simulations per sample size}

\item{parallel}{logical}

\item{ncores}{number of cores to use in parallel computation}
}
\value{
a dataframe
}
\description{
Estimate power curve for Bland-Altman limits of agreement test
}
