% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePower.R
\name{estimatePower}
\alias{estimatePower}
\title{Estimate power for Bland-Altman limits of agreement test}
\usage{
estimatePower(powerCurve, n)
}
\arguments{
\item{powerCurve}{an object produced by estimatePowerCurve}

\item{power}{the pre-determined power}
}
\value{
a list
}
\description{
Estimate power for Bland-Altman limits of agreement test
}
